requirements.txt.txt
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.1.0